refactor: extract non-interactive mode logic into separate file
